seasonal/season_2_feb_2025/draft/download-zoom-recordings-appscript/python_clone/3_more/requirements.txt
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
requests==2.31.0
tqdm==4.66.1
loguru==0.7.2 